User auth
Added users with the ability to log in and out.

Users are stored as an array in $app['users'] with the key being the username and the value a bcrypted password.

Moved templates and content into an app directory.
Added a redirect and flash service provider that help with user auth stuff.

Removed old views.